Make command execution error handling more fine-grained
execute_command does not throw an exception when an
error occurs if argument ignore_errors is true. However,
it still prints a warning message in this case.

Introduce a new option silent_errors that allows to even switch off this
message (in case a failing command is expected and does
not need to cause a warning message).

Of course, this option should be used with caution.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>