Refactor git log parsing in the Conway analysis
This commit introduces two changes to the git log parsing
mechanisms: First, make sure that empty commits are handled
properly. Such sommits are not separated by \n\n in the
output of git log --numstat, and this is now appropriately
handled in the parser. Second, reduce the amount of output
fields in the output of parseGitLogOutput() by using
single fields for dates (date, hour and timezone are placed
in one single column). Also make sure in the subsequent code
to use proper date information.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>