Properly compute mail ID maps per revision range
To distinguish between local mail IDs (mails are consecutively
numbered from 1 to N for each revision range when we process
them in the corpus) and the in-DB IDs (assigned globally
unique by the DB), we compute a simple map between these
numbers. Since the local IDs overlap across release ranges,
the map of course needs to be computed locally for each
range. Extend the DB query (and the caller) accordingly.

Since double IDs can happen in error cases, add a new parameter
to ml.thread.loc.to.glob that allows for returning all IDs
in such cases.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>