Ensure that results from compute.ee.relations don't contain factors
When factors are used for the entities (V1 and V2) in the
dependency data frame, interleave operations will result
in a list of interleaved numbers instead of the intended
names (like filenames). This introduces bogus edges into
the person-artefact graph.

Use the opportunity to add some comments to the function.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>