Refactor query.dependency parameter passing
The function receives a huge list of parameters, but many
of the values are available in the conf object. Use this fact,
and also equip some of the dependent functions with shorter
argument lists. These changes set the stage to extend the
analysis to multiple motif types that are computed
sequentially.

Add some documentation for the various analysis parameters
as by-catch.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>