Fix conway correlation vs. raw data naming mess
Once the conway covariables are extracted from the raw data, we use them
(amongst others) for two purposes: Computing correlations, and computing
time series. The current state of the code mixes these two concepts by
naming functions that prepare raw data (that can be used
for time series and correlation calculations) along the lines
of ***.correlation.***. This is confusing -- distinguish between
both concepts, and unclutter the code in the process.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>