Conway analysis: Construct range-specific filenames only when they are needed
Filenames for range-specific files require two pieces of information:
The name of the file, and the path where it resides. Since the latter
is anyway known when such files are read, and the file name as
such is a project-global convention (that will additionally go
away once the data end up in the database proper), we can construct
the full filenames when they are actually needed, and not push
them around via function parameters.

This change is the basis for combining results from different time
windows to check if there are long-range temporal conway correlations.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>