Move quality analysis into global conway pass
When computing correlations between different release ranges
in the local pass, we face the problem that computing correlations
between cycles i and i+1 requires data for cycle i+1 to be already
available, which they are not when the cycles are iterated through
sequentially. Instead of implementing involved iterators,
move the correlation analysis into the global pass.

Use the opportunity for some spring cleaning.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>