Bring conway.py closer to sanity
Parsing and processing jira issues currently imports libraries pandes
and numpy for the sake of 5 lines of python. Rewrite using 6 lines
of standard python, and save about 50% of the installation dependencies
of Codeface.

While at it, improve variable naming, add some comments, and
fix indentation. Additionally, also include cretor entries
of issues in the communication list (and not just comments).

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>