Ignore problems with old urllib versions
While there are known issues with https downloads via old
urllib versions, these are not relevant for us since we're
only grabbing public data, and don't need to keep the content
secret. While it is usually of course not adivsable to ignore
SSL warnings, doing so won't do harm in this case. It eliminates
tons of warning messages, though, that make the progessbar for
jira email resolution unusable.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>